Move ScheduledWork and DecoderWorker to the .cc file. Make temp_buffers a part of ScheduledWork and clear them after the work is done.
