+++ dali/imgcodec/image_decoder.cc
@@ -12,8 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#include <cassert>
+#include <condition_variable>
 #include <memory>
 #include <map>
+#include <mutex>
 #include <vector>
 #include "dali/core/cuda_error.h"
 #include "dali/core/mm/memory.h"
@@ -22,6 +25,350 @@
 
 namespace dali {
 namespace imgcodec {
+
+///////////////////////////////////////////////////////////////////////////////
+// ScheduledWork
+///////////////////////////////////////////////////////////////////////////////
+
+/**
+ * @brief Describes a sub-batch of work to be processed
+ *
+ * This object contains a (shared) promise from the original request containing the full batch
+ * and a mapping of indices from this sub-batch to the full batch.
+ * It also contains a subset of relevant sample views, sources, etc.
+ */
+struct ImageDecoder::ScheduledWork {
+  ScheduledWork(DecodeContext ctx, DecodeResultsPromise results)
+  : ctx(std::move(ctx)), results(std::move(results)) {}
+
+  void clear() {
+    indices.clear();
+    sources.clear();
+    cpu_outputs.clear();
+    gpu_outputs.clear();
+    rois.clear();
+    temp_buffers.clear();
+  }
+
+  int num_samples() const {
+    return indices.size();
+  }
+
+  bool empty() const {
+    return indices.empty();
+  }
+
+  void resize(int num_samples) {
+    indices.resize(num_samples);
+    sources.resize(num_samples);
+    if (!cpu_outputs.empty())
+      cpu_outputs.resize(num_samples);
+    if (!gpu_outputs.empty())
+      gpu_outputs.resize(num_samples);
+    if (!rois.empty())
+      rois.resize(num_samples);
+    if (!temp_buffers.empty())
+      temp_buffers.resize(num_samples);
+  }
+
+  /**
+   * @brief Moves one work entry from another work to this one
+   */
+  void move_entry(ScheduledWork &from, int which);
+
+  /**
+   * @brief Allocates temporary CPU outputs for this sub-batch
+   *
+   * This function is used when falling back from GPU to CPU decoder.
+   */
+  void alloc_temp_cpu_outputs();
+
+  DecodeContext ctx;
+  // The original promise
+  DecodeResultsPromise results;
+  // The indices in the original request
+  BatchVector<int> indices;
+  BatchVector<ImageSource *> sources;
+  BatchVector<SampleView<CPUBackend>> cpu_outputs;
+  BatchVector<SampleView<GPUBackend>> gpu_outputs;
+  BatchVector<mm::async_uptr<void>> temp_buffers;
+  DecodeParams params;
+  BatchVector<ROI> rois;
+
+  std::unique_ptr<ScheduledWork> next;
+};
+
+void ImageDecoder::ScheduledWork::move_entry(ScheduledWork &from, int which) {
+  sources.push_back(from.sources[which]);
+  indices.push_back(from.indices[which]);
+  if (!from.cpu_outputs.empty())
+    cpu_outputs.push_back(std::move(from.cpu_outputs[which]));
+  if (!from.gpu_outputs.empty())
+    gpu_outputs.push_back(std::move(from.gpu_outputs[which]));
+  if (!from.temp_buffers.empty())
+    temp_buffers.push_back(std::move(from.temp_buffers[which]));
+  if (!from.rois.empty())
+    rois.push_back(std::move(from.rois[which]));
+}
+
+void ImageDecoder::ScheduledWork::alloc_temp_cpu_outputs() {
+  cpu_outputs.resize(indices.size());
+  temp_buffers.clear();
+  for (int i = 0, n = indices.size(); i < n; i++) {
+    SampleView<GPUBackend> &gpu_out = gpu_outputs[i];
+
+    // TODO(michalz): Add missing utility functions to SampleView - or just use Tensor again...
+    size_t size = volume(gpu_out.shape()) * TypeTable::GetTypeInfo(gpu_out.type()).size();
+    constexpr int kTempBufferAlignment = 256;
+
+    temp_buffers.emplace_back();
+    auto &buf_ptr = temp_buffers.back();
+    buf_ptr = mm::alloc_raw_async_unique<char, mm::memory_kind::pinned>(
+      size, mm::host_sync, ctx.stream, kTempBufferAlignment);
+
+    SampleView<CPUBackend> &cpu_out = cpu_outputs[i];
+    cpu_out = SampleView<CPUBackend>(buf_ptr.get(), gpu_out.shape(), gpu_out.type());
+  }
+}
+
+
+///////////////////////////////////////////////////////////////////////////////
+// DecoderWorker
+///////////////////////////////////////////////////////////////////////////////
+
+/**
+ * @brief A worker that processes sub-batches of work to be processed by a particular decoder.
+ *
+ * A DecoderWorker waits for incoming ScheduledWork objects and processes them by running
+ * `decoder_->ScheduleDecode` and waiting for partial results, scheduling the failed
+ * samples to a fallback decoder, if present.
+ *
+ * When a sample is successfully decoded, it is marked as a success in the parent
+ * DecodeResultsPromise. If it fails, it goes to fallback and only if all fallbacks fail, it is
+ * marked in the DecodeResultsPromise as a failure.
+ */
+class ImageDecoder::DecoderWorker {
+ public:
+  /**
+   * @brief Constructs a decoder worker for a given decoder.
+   *
+   * @param owner   - the parent decoder object
+   * @param factory - the factory that constructs the decoder for this worker
+   * @param start   - if true, the decoder is immediately instantiated and the worker thread
+   *                  is launched; otherwise a call to `start` is delayed until the first
+   *                  work that's relevant for this decoder.
+   */
+  DecoderWorker(ImageDecoder *owner, const ImageDecoderFactory *factory, bool start) {
+    owner_ = owner;
+    factory_ = factory;
+    if (start)
+      this->start();
+  }
+  ~DecoderWorker();
+
+  void start();
+  void stop();
+  void add_work(std::unique_ptr<ScheduledWork> work);
+
+  void set_fallback(DecoderWorker *fallback) {
+    fallback_ = fallback;
+  }
+
+  ImageDecoderInstance *decoder(bool create_if_null = true);
+
+ private:
+  std::mutex mtx_;
+  std::condition_variable cv_;
+
+  std::unique_ptr<ScheduledWork> work_;
+  std::thread worker_;
+  bool stop_requested_ = false;
+  std::once_flag started_;
+
+  ImageDecoder *owner_ = nullptr;
+  const ImageDecoderFactory *factory_ = nullptr;
+  std::shared_ptr<ImageDecoderInstance> decoder_;
+  bool produces_gpu_output_ = false;
+
+  /**
+   * @brief Processes a (sub)batch of work.
+   *
+   * The work is scheduled and the results are waited for. Any failed samples will be added
+   * to a fallback work, if a fallback decoder is present.
+   */
+  void process_batch(std::unique_ptr<ScheduledWork> work) noexcept;
+
+  /**
+   * @brief The main loop of the worker thread.
+   */
+  void run();
+
+  // Fallback worker or null, if no fallback is available
+  DecoderWorker *fallback_ = nullptr;
+};
+
+ImageDecoderInstance *ImageDecoder::DecoderWorker::decoder(bool create_if_null) {
+  if (!decoder_) {
+    decoder_ = factory_->Create(owner_->device_id_, owner_->params_);
+    produces_gpu_output_ = factory_->GetProperties().gpu_output;
+  }
+  return decoder_.get();
+}
+
+
+void ImageDecoder::DecoderWorker::start() {
+  std::call_once(started_, [&]() {
+    (void)decoder(true);
+    worker_ = std::thread(&DecoderWorker::run, this);
+  });
+}
+
+ImageDecoder::DecoderWorker::~DecoderWorker() {
+  stop();
+}
+
+void ImageDecoder::DecoderWorker::stop() {
+  if (worker_.joinable()) {
+    {
+      std::lock_guard lock(mtx_);
+      stop_requested_ = true;
+      work_.reset();
+    }
+    cv_.notify_all();
+    worker_.join();
+    worker_ = {};
+  }
+}
+
+void ImageDecoder::DecoderWorker::run() {
+  DeviceGuard dg(owner_->device_id_);
+  while (!stop_requested_) {
+    std::unique_lock lock(mtx_);
+    cv_.wait(lock, [&]() {
+      return stop_requested_ || work_ != nullptr;
+    });
+    if (stop_requested_)
+      break;
+    assert(work_ != nullptr);
+    auto w = std::move(work_);
+    lock.unlock();
+    process_batch(std::move(w));
+  }
+}
+
+void ImageDecoder::DecoderWorker::add_work(std::unique_ptr<ScheduledWork> work) {
+  assert(work->num_samples() > 0);
+  {
+    std::lock_guard guard(mtx_);
+    assert((work->cpu_outputs.empty() && work->gpu_outputs.size() == work->sources.size()) ||
+           (work->gpu_outputs.empty() && work->cpu_outputs.size() == work->sources.size()));
+    assert(work->rois.empty() || work->rois.size() == work->sources.size());
+    if (work_) {
+      owner_->combine_work(*work_, std::move(work));
+      // no need to notify - a work item was already there, so it will be picked up regardless
+    } else {
+      work_ = std::move(work);
+      cv_.notify_one();
+    }
+  }
+  start();
+}
+
+// The main processing function
+void ImageDecoder::DecoderWorker::process_batch(std::unique_ptr<ScheduledWork> work) noexcept {
+  assert(work->num_samples() > 0);
+  assert((work->cpu_outputs.empty() && work->gpu_outputs.size() == work->sources.size()) ||
+         (work->gpu_outputs.empty() && work->cpu_outputs.size() == work->sources.size()));
+  assert(work->rois.empty() || work->rois.size() == work->sources.size());
+
+  auto mask = decoder_->CanDecode(work->ctx,
+                                  make_span(work->sources),
+                                  work->params,
+                                  make_span(work->rois));
+
+  std::unique_ptr<ScheduledWork> fallback_work;
+  if (fallback_) {
+    fallback_work = owner_->new_work(work->ctx, work->results, work->params);
+    move_to_fallback(fallback_work.get(), *work, mask);
+    if (!fallback_work->empty())
+      fallback_->add_work(std::move(fallback_work));
+  } else {
+    filter(*work, mask);
+    for (size_t i = 0; i < mask.size(); i++) {
+      if (!mask[i])
+        work->results.set(work->indices[i], DecodeResult::Failure(nullptr));
+    }
+  }
+
+  if (!work->sources.empty()) {
+    bool decode_to_gpu = produces_gpu_output_;
+
+    if (!decode_to_gpu && work->cpu_outputs.empty()) {
+      work->alloc_temp_cpu_outputs();
+    }
+
+    auto future = decode_to_gpu
+      ? decoder_->ScheduleDecode(work->ctx,
+                                 make_span(work->gpu_outputs),
+                                 make_span(work->sources),
+                                 work->params,
+                                 make_span(work->rois))
+      : decoder_->ScheduleDecode(work->ctx,
+                                 make_span(work->cpu_outputs),
+                                 make_span(work->sources),
+                                 work->params,
+                                 make_span(work->rois));
+
+    if (fallback_) {
+      for (;;) {
+        auto indices = future.wait_new();
+        if (indices.empty())
+          break;
+
+        auto r = future.get_all_ref();
+
+        for (int sub_idx : indices) {
+          if (r[sub_idx].success) {
+            work->results.set(work->indices[sub_idx], r[sub_idx]);
+            if (!decode_to_gpu && !work->gpu_outputs.empty()) {
+              copy(work->gpu_outputs[sub_idx], work->cpu_outputs[sub_idx], work->ctx.stream);
+            }
+          } else {
+            if (!fallback_work)
+              fallback_work = owner_->new_work(work->ctx, work->results, work->params);
+            fallback_work->move_entry(*work, sub_idx);
+          }
+        }
+
+        if (fallback_work && !fallback_work->empty())
+          fallback_->add_work(std::move(fallback_work));
+      }
+    } else {
+      // no fallback - the results are final
+      future.wait_all();
+      auto r = future.get_all();
+      // we may need to copy to the gpu
+      if (!decode_to_gpu && !work->gpu_outputs.empty()) {
+        for (int i = 0; i < r.size(); i++) {
+          if (r[i].success) {  // TODO(michalz): bulk copy
+            copy(work->gpu_outputs[i], work->cpu_outputs[i], work->ctx.stream);
+          }
+        }
+      }
+      // just copy the result to the final promise
+      for (int i = 0; i < work->num_samples(); i++) {
+        work->results.set(work->indices[i], r[i]);
+      }
+    }
+  }
+  owner_->recycle_work(std::move(work));
+}
+
+
+///////////////////////////////////////////////////////////////////////////////
+// ImageDecoder
+///////////////////////////////////////////////////////////////////////////////
+
 
 ImageDecoder::ImageDecoder(int device_id,
                            bool lazy_init,
@@ -197,13 +544,12 @@ FutureDecodeResults ImageDecoder::ScheduleDecode(DecodeContext ctx,
                                                 cspan<ImageSource *> in,
                                                 DecodeParams opts,
                                                 cspan<ROI> rois) {
-  temp_buffers_.clear();
   int N = out.size();
   assert(in.size() == N);
   assert(rois.size() == N || rois.empty());
   DecodeResultsPromise results(N);
 
-  auto work = new_work(ctx, results);
+  auto work = new_work(ctx, results, opts);
   work->cpu_outputs.reserve(N);
   work->indices.reserve(N);
   work->sources.reserve(N);
@@ -216,8 +562,6 @@ FutureDecodeResults ImageDecoder::ScheduleDecode(DecodeContext ctx,
     work->indices.push_back(i);
 
   DistributeWork(std::move(work));
-
-  temp_buffers_.clear();
 
   return results.get_future();
 }
@@ -235,11 +579,9 @@ FutureDecodeResults ImageDecoder::ScheduleDecode(DecodeContext ctx,
   assert(in.size() == N);
   assert(rois.size() == N || rois.empty());
 
-  temp_buffers_.resize(N);
-
   DecodeResultsPromise results(N);
 
-  auto work = new_work(ctx, results);
+  auto work = new_work(ctx, results, opts);
   work->gpu_outputs.reserve(N);
   work->indices.reserve(N);
   work->sources.reserve(N);
@@ -273,7 +615,7 @@ void ImageDecoder::DistributeWork(std::unique_ptr<ScheduledWork> work) {
     }
     auto &w = dist[f];
     if (!w)
-      w = new_work(work->ctx, work->results);
+      w = new_work(work->ctx, work->results, work->params);
     w->move_entry(*work, i);
   }
   for (auto &[f, w] : dist) {
@@ -304,55 +646,6 @@ void ImageDecoder::InitWorkers(bool lazy_init) {
   }
 }
 
-ImageDecoderInstance *ImageDecoder::DecoderWorker::decoder(bool create_if_null) {
-  if (!decoder_) {
-    decoder_ = factory_->Create(owner_->device_id_, owner_->params_);
-    produces_gpu_output_ = factory_->GetProperties().gpu_output;
-  }
-  return decoder_.get();
-}
-
-
-void ImageDecoder::DecoderWorker::start() {
-  std::call_once(started_, [&]() {
-    (void)decoder(true);
-    worker_ = std::thread(&DecoderWorker::run, this);
-  });
-}
-
-ImageDecoder::DecoderWorker::~DecoderWorker() {
-  stop();
-}
-
-void ImageDecoder::DecoderWorker::stop() {
-  if (worker_.joinable()) {
-    {
-      std::lock_guard lock(mtx_);
-      stop_requested_ = true;
-      work_.reset();
-    }
-    cv_.notify_all();
-    worker_.join();
-    worker_ = {};
-  }
-}
-
-void ImageDecoder::DecoderWorker::run() {
-  DeviceGuard dg(owner_->device_id_);
-  while (!stop_requested_) {
-    std::unique_lock lock(mtx_);
-    cv_.wait(lock, [&]() {
-      return stop_requested_ || work_ != nullptr;
-    });
-    if (stop_requested_)
-      break;
-    assert(work_ != nullptr);
-    auto w = std::move(work_);
-    lock.unlock();
-    process_batch(std::move(w));
-  }
-}
-
 void ImageDecoder::move_to_fallback(ScheduledWork *fb,
                                     ScheduledWork &work,
                                     const vector<bool> &keep) {
@@ -368,6 +661,8 @@ void ImageDecoder::move_to_fallback(ScheduledWork *fb,
             work.cpu_outputs[i - moved] = std::move(work.cpu_outputs[i]);
           if (!work.gpu_outputs.empty())
             work.gpu_outputs[i - moved] = std::move(work.gpu_outputs[i]);
+          if (!work.temp_buffers.empty())
+            work.temp_buffers[i - moved] = std::move(work.temp_buffers[i]);
           if (!work.rois[i])
             work.rois[i - moved] = std::move(work.rois[i]);
           work.sources[i - moved] = work.sources[i];
@@ -383,114 +678,6 @@ void ImageDecoder::move_to_fallback(ScheduledWork *fb,
       fb->resize(fb->num_samples() - moved);
 }
 
-void ImageDecoder::ScheduledWork::alloc_temp_cpu_outputs(ImageDecoder &owner) {
-  cpu_outputs.resize(indices.size());
-  for (int i = 0, n = indices.size(); i < n; i++) {
-    int idx = indices[i];
-
-    SampleView<GPUBackend> &gpu_out = gpu_outputs[i];
-
-    // TODO(michalz): Add missing utility functions to SampleView - or just use Tensor again...
-    size_t size = volume(gpu_out.shape()) * TypeTable::GetTypeInfo(gpu_out.type()).size();
-    constexpr int kTempBufferAlignment = 256;
-
-    auto &buf_ptr = owner.temp_buffers_[idx];
-    buf_ptr = mm::alloc_raw_async_shared<char, mm::memory_kind::pinned>(
-      size, mm::host_sync, ctx.stream, kTempBufferAlignment);
-
-    SampleView<CPUBackend> &cpu_out = cpu_outputs[i];
-    cpu_out = SampleView<CPUBackend>(buf_ptr.get(), gpu_out.shape(), gpu_out.type());
-  }
-}
-
-void ImageDecoder::DecoderWorker::process_batch(std::unique_ptr<ScheduledWork> work) noexcept {
-  assert(work->num_samples() > 0);
-  assert((work->cpu_outputs.empty() && work->gpu_outputs.size() == work->sources.size()) ||
-         (work->gpu_outputs.empty() && work->cpu_outputs.size() == work->sources.size()));
-  assert(work->rois.empty() || work->rois.size() == work->sources.size());
-
-  auto mask = decoder_->CanDecode(work->ctx,
-                                  make_span(work->sources),
-                                  work->params,
-                                  make_span(work->rois));
-
-  std::unique_ptr<ScheduledWork> fallback_work;
-  if (fallback_) {
-    fallback_work = owner_->new_work(work->ctx, work->results);
-    move_to_fallback(fallback_work.get(), *work, mask);
-    if (!fallback_work->empty())
-      fallback_->add_work(std::move(fallback_work));
-  } else {
-    filter(*work, mask);
-    for (size_t i = 0; i < mask.size(); i++) {
-      if (!mask[i])
-        work->results.set(work->indices[i], DecodeResult::Failure(nullptr));
-    }
-  }
-
-  if (!work->sources.empty()) {
-    bool decode_to_gpu = produces_gpu_output_;
-
-    if (!decode_to_gpu && work->cpu_outputs.empty()) {
-      work->alloc_temp_cpu_outputs(*owner_);
-    }
-
-    auto future = decode_to_gpu
-      ? decoder_->ScheduleDecode(work->ctx,
-                                 make_span(work->gpu_outputs),
-                                 make_span(work->sources),
-                                 work->params,
-                                 make_span(work->rois))
-      : decoder_->ScheduleDecode(work->ctx,
-                                 make_span(work->cpu_outputs),
-                                 make_span(work->sources),
-                                 work->params,
-                                 make_span(work->rois));
-
-    if (fallback_) {
-      for (;;) {
-        auto indices = future.wait_new();
-        if (indices.empty())
-          break;
-
-        auto r = future.get_all_ref();
-
-        for (int sub_idx : indices) {
-          if (r[sub_idx].success) {
-            work->results.set(work->indices[sub_idx], r[sub_idx]);
-            if (!decode_to_gpu && !work->gpu_outputs.empty()) {
-              copy(work->gpu_outputs[sub_idx], work->cpu_outputs[sub_idx], work->ctx.stream);
-            }
-          } else {
-            if (!fallback_work)
-              fallback_work = owner_->new_work(work->ctx, work->results);
-            fallback_work->move_entry(*work, sub_idx);
-          }
-        }
-
-        if (fallback_work && !fallback_work->empty())
-          fallback_->add_work(std::move(fallback_work));
-      }
-    } else {
-      // no fallback - the results are final
-      future.wait_all();
-      auto r = future.get_all();
-      // we may need to copy to the gpu
-      if (!decode_to_gpu && !work->gpu_outputs.empty()) {
-        for (int i = 0; i < r.size(); i++) {
-          if (r[i].success) {  // TODO(michalz): bulk copy
-            copy(work->gpu_outputs[i], work->cpu_outputs[i], work->ctx.stream);
-          }
-        }
-      }
-      // just copy the result to the final promise
-      for (int i = 0; i < work->num_samples(); i++) {
-        work->results.set(work->indices[i], r[i]);
-      }
-    }
-  }
-  owner_->recycle_work(std::move(work));
-}
 
 void ImageDecoder::copy(SampleView<GPUBackend> &out,
                         const ConstSampleView<CPUBackend> &in,
@@ -504,25 +691,52 @@ void ImageDecoder::copy(SampleView<GPUBackend> &out,
                             cudaMemcpyHostToDevice, stream));
 }
 
+std::vector<DecodeResult> ImageDecoder::Decode(DecodeContext ctx,
+                                               TensorList<CPUBackend> &out,
+                                               cspan<ImageSource *> in,
+                                               DecodeParams opts,
+                                               cspan<ROI> rois) {
+  BatchVector<SampleView<CPUBackend>> samples;
+  int N = out.num_samples();
+  samples.resize(N);
+  for (int i = 0; i < N; i++)
+    samples[i] = out[i];
+  return Decode(ctx, make_span(samples), in, opts, rois);
+}
+
+std::vector<DecodeResult> ImageDecoder::Decode(DecodeContext ctx,
+                                               TensorList<GPUBackend> &out,
+                                               cspan<ImageSource *> in,
+                                               DecodeParams opts,
+                                               cspan<ROI> rois) {
+  BatchVector<SampleView<GPUBackend>> samples;
+  int N = out.num_samples();
+  samples.resize(N);
+  for (int i = 0; i < N; i++)
+    samples[i] = out[i];
+  return Decode(ctx, make_span(samples), in, opts, rois);
+}
+
 ////////////////////////////////////////////////////////////////////////////
 // Work item management
+////////////////////////////////////////////////////////////////////////////
 
 std::unique_ptr<ImageDecoder::ScheduledWork> ImageDecoder::new_work(
-    DecodeContext ctx, DecodeResultsPromise results) {
+    DecodeContext ctx, DecodeResultsPromise results, DecodeParams params) {
   if (free_work_items_) {
     std::lock_guard<std::mutex> g(work_mutex_);
     if (free_work_items_) {
       auto ptr = std::move(free_work_items_);
       free_work_items_ = std::move(ptr->next);
       ptr->results = std::move(results);
-      ptr->ctx = ctx;
+      ptr->ctx = std::move(ctx);
+      ptr->params = std::move(params);
       return ptr;
     }
   }
 
   return std::make_unique<ScheduledWork>(ctx, results);
 }
-
 
 void ImageDecoder::recycle_work(std::unique_ptr<ScheduledWork> work) {
   std::lock_guard<std::mutex> g(work_mutex_);
@@ -553,62 +767,6 @@ void ImageDecoder::combine_work(ScheduledWork &target, std::unique_ptr<Scheduled
   move_append(target.rois, source->rois);
   recycle_work(std::move(source));
 }
-
-void ImageDecoder::ScheduledWork::move_entry(ScheduledWork &from, int which) {
-  sources.push_back(from.sources[which]);
-  indices.push_back(from.indices[which]);
-  if (!from.cpu_outputs.empty())
-    cpu_outputs.push_back(std::move(from.cpu_outputs[which]));
-  if (!from.gpu_outputs.empty())
-    gpu_outputs.push_back(std::move(from.gpu_outputs[which]));
-  if (!from.rois.empty())
-    rois.push_back(std::move(from.rois[which]));
-}
-
-void ImageDecoder::DecoderWorker::add_work(std::unique_ptr<ScheduledWork> work) {
-  assert(work->num_samples() > 0);
-  {
-    std::lock_guard guard(mtx_);
-    assert((work->cpu_outputs.empty() && work->gpu_outputs.size() == work->sources.size()) ||
-           (work->gpu_outputs.empty() && work->cpu_outputs.size() == work->sources.size()));
-    assert(work->rois.empty() || work->rois.size() == work->sources.size());
-    if (work_) {
-      owner_->combine_work(*work_, std::move(work));
-      // no need to notify - a work item was already there, so it will be picked up regardless
-    } else {
-      work_ = std::move(work);
-      cv_.notify_one();
-    }
-  }
-  start();
-}
-
-std::vector<DecodeResult> ImageDecoder::Decode(DecodeContext ctx,
-                                               TensorList<CPUBackend> &out,
-                                               cspan<ImageSource *> in,
-                                               DecodeParams opts,
-                                               cspan<ROI> rois) {
-  BatchVector<SampleView<CPUBackend>> samples;
-  int N = out.num_samples();
-  samples.resize(N);
-  for (int i = 0; i < N; i++)
-    samples[i] = out[i];
-  return Decode(ctx, make_span(samples), in, opts, rois);
-}
-
-std::vector<DecodeResult> ImageDecoder::Decode(DecodeContext ctx,
-                                               TensorList<GPUBackend> &out,
-                                               cspan<ImageSource *> in,
-                                               DecodeParams opts,
-                                               cspan<ROI> rois) {
-  BatchVector<SampleView<GPUBackend>> samples;
-  int N = out.num_samples();
-  samples.resize(N);
-  for (int i = 0; i < N; i++)
-    samples[i] = out[i];
-  return Decode(ctx, make_span(samples), in, opts, rois);
-}
-
 
 }  // namespace imgcodec
 }  // namespace dali
