+++ include/dali/imgcodec/image_decoder.h
@@ -15,7 +15,6 @@
 #ifndef DALI_IMGCODEC_IMAGE_DECODER_H_
 #define DALI_IMGCODEC_IMAGE_DECODER_H_
 
-#include <condition_variable>
 #include <map>
 #include <memory>
 #include <mutex>
@@ -187,76 +186,15 @@ class DLL_PUBLIC ImageDecoder : public ImageDecoderInstance, public ImageParser 
   any GetParam(const char *key) const override;
 
  private:
-  /**
-   * @brief Describes a sub-batch of work to be processed
-   *
-   * This object contains a (shared) promise from the original request containing the full batch
-   * and a mapping of indices from this sub-batch to the full batch.
-   * It also contains a subset of relevant sample views, sources, etc.
-   */
-  struct ScheduledWork {
-    ScheduledWork(DecodeContext ctx, DecodeResultsPromise results)
-    : ctx(std::move(ctx)), results(std::move(results)) {}
-
-    void clear() {
-      indices.clear();
-      sources.clear();
-      cpu_outputs.clear();
-      gpu_outputs.clear();
-      rois.clear();
-    }
-
-    int num_samples() const {
-      return indices.size();
-    }
-
-    bool empty() const {
-      return indices.empty();
-    }
-
-    void resize(int num_samples) {
-      indices.resize(num_samples);
-      sources.resize(num_samples);
-      if (!cpu_outputs.empty())
-        cpu_outputs.resize(num_samples);
-      if (!gpu_outputs.empty())
-        gpu_outputs.resize(num_samples);
-      if (!rois.empty())
-        rois.resize(num_samples);
-    }
-
-    /**
-     * @brief Moves one work entry from another work to this one
-     */
-    void move_entry(ScheduledWork &from, int which);
-
-    /**
-     * @brief Allocates temporary CPU outputs for this sub-batch
-     *
-     * This function is used when falling back from GPU to CPU decoder.
-     */
-    void alloc_temp_cpu_outputs(ImageDecoder &owner);
-
-    DecodeContext ctx;
-    // The original promise
-    DecodeResultsPromise results;
-    // The indices in the original request
-    BatchVector<int> indices;
-    BatchVector<ImageSource *> sources;
-    BatchVector<SampleView<CPUBackend>> cpu_outputs;
-    BatchVector<SampleView<GPUBackend>> gpu_outputs;
-    DecodeParams params;
-    BatchVector<ROI> rois;
-
-    std::unique_ptr<ScheduledWork> next;
-  };
-
+  struct ScheduledWork;
   // Scheduled work pool
 
   std::mutex work_mutex_;
   std::unique_ptr<ScheduledWork> free_work_items_;
 
-  std::unique_ptr<ScheduledWork> new_work(DecodeContext ctx, DecodeResultsPromise results);
+  std::unique_ptr<ScheduledWork> new_work(DecodeContext ctx,
+                                          DecodeResultsPromise results,
+                                          DecodeParams params);
 
   void recycle_work(std::unique_ptr<ScheduledWork> work);
 
@@ -270,80 +208,16 @@ class DLL_PUBLIC ImageDecoder : public ImageDecoderInstance, public ImageParser 
     move_to_fallback(nullptr, work, keep);
   }
 
-  /**
-   * @brief A worker that processes sub-batches of work to be processed by a particular decoder.
-   *
-   * A DecoderWorker waits for incoming ScheduledWork objects and processes them by running
-   * `decoder_->ScheduleDecode` and waiting for partial results, scheduling the failed
-   * samples to a fallback decoder, if present.
-   *
-   * When a sample is successfully decoded, it is marked as succeess in the parent
-   * DecodeResultsPromise. If it fails, it goes to fallback and only if all fallbacks fail, it is
-   * marked in the DecodeResultsPromise as a failure.
-   */
-  class DecoderWorker {
-   public:
-    /**
-     * @brief Constructs a decoder worker for a given decoder.
-     *
-     * @param owner   - the parent decoder object
-     * @param factory - the factory that constructs the decoder for this worker
-     * @param start   - if true, the decoder is immediately instantiated and the worker thread
-     *                  is launched; otherwise a call to `start` is delayed until the first
-     *                  work that's relevant for this decoder.
-     */
-    DecoderWorker(ImageDecoder *owner, const ImageDecoderFactory *factory, bool start) {
-      owner_ = owner;
-      factory_ = factory;
-      if (start)
-        this->start();
-    }
-    ~DecoderWorker();
+  void alloc_temp_cpu_outputs(ScheduledWork &work);
 
-    void start();
-    void stop();
-    void add_work(std::unique_ptr<ScheduledWork> work);
-
-    void set_fallback(DecoderWorker *fallback) {
-      fallback_ = fallback;
-    }
-
-    ImageDecoderInstance *decoder(bool create_if_null = true);
-
-   private:
-    std::mutex mtx_;
-    std::condition_variable cv_;
-
-    std::unique_ptr<ScheduledWork> work_;
-    std::thread worker_;
-    bool stop_requested_ = false;
-    std::once_flag started_;
-
-    ImageDecoder *owner_ = nullptr;
-    const ImageDecoderFactory *factory_ = nullptr;
-    std::shared_ptr<ImageDecoderInstance> decoder_;
-    bool produces_gpu_output_ = false;
-
-    /**
-     * @brief Processes a (sub)batch of work.
-     *
-     * The work is scheduled and the results are waited for. Any failed samples will be added
-     * to a fallback work, if a fallback decoder is present.
-     */
-    void process_batch(std::unique_ptr<ScheduledWork> work) noexcept;
-
-    /**
-     * @brief The main loop of the worker thread.
-     */
-    void run();
-
-    // Fallback worker or null, if no fallback is available
-    DecoderWorker *fallback_ = nullptr;
-  };
+  // Decoder workers
+  class DecoderWorker;
 
   void InitWorkers(bool lazy);
 
   void DistributeWork(std::unique_ptr<ScheduledWork> work);
+
+  // Miscellanous
 
   static void copy(SampleView<GPUBackend> &out,
                    const ConstSampleView<CPUBackend> &in,
@@ -353,7 +227,6 @@ class DLL_PUBLIC ImageDecoder : public ImageDecoderInstance, public ImageParser 
   std::map<std::string, any> params_;
   std::map<const ImageDecoderFactory*, std::unique_ptr<DecoderWorker>> workers_;
   std::multimap<const ImageFormat*, ImageDecoderFactory*> filtered_;
-  std::vector<std::shared_ptr<void>> temp_buffers_;
   struct TempImageSource {
     // TODO(michalz): store auxiliary image sources here
   };
